Move log exception methods along with their associated helpers into the abstraction layer.
