+++ Com.Kumulos.Android/KumulosImplementation.cs
@@ -205,15 +205,7 @@ namespace Com.Kumulos
             Android.Kumulos.ClearUserAssociation(Application.Context.ApplicationContext);
         }
 
-        public void LogException(Exception e)
-        {
-            AttemptToLogException(e, false);
-        }
-
-        public void LogUncaughtException(Exception e)
-        {
-            AttemptToLogException(e, true);
-        }
+      
 
         public void SendLocationUpdate(double lat, double lng)
         {
@@ -230,40 +222,7 @@ namespace Com.Kumulos
             Android.Kumulos.TrackEddystoneBeaconProximity(Application.Context.ApplicationContext, namespaceHex, instanceHex, dblDistance);
         }
 
-        private void AttemptToLogException(Exception e, bool uncaught)
-        {
-            try
-            {
-                var dict = GetDictionaryForException(e, uncaught);
-                WriteCrashToDisk(dict);
-            }
-            catch (Exception ex)
-            {
-                //- Don't cause further exceptions trying to log exceptions.
-            }
-        }
-
-        private Dictionary<string, object> GetDictionaryForException(Exception e, bool uncaught)
-        {
-            var st = new StackTrace(e, true);
-            var frame = st.GetFrame(0);
-            var line = frame.GetFileLineNumber();
-
-            var dict = GetDictionaryForExceptionTracking(e, uncaught);
-
-            var report = (Dictionary<string, object>)dict["report"];
-            report.Add("lineNumber", line);
-
-            return dict;
-        }
-
-        private void WriteCrashToDisk(Dictionary<string, object> crash)
-        {
-            var documents = Environment.GetFolderPath(Environment.SpecialFolder.Personal);
-            var filename = Path.Combine(documents, "CrashLog.json");
-            File.WriteAllText(filename, JsonConvert.SerializeObject(crash, Formatting.None));
-        }
-
+      
        
 
         public void TrackiBeaconProximity(object CLBeaconObject)
